+++ src/derivatives_trading_usds_futures/websocket_api/apis/trade_api.rs
@@ -21,6 +21,7 @@ use serde_json::Value;
 use std::{collections::BTreeMap, sync::Arc};
 
 use crate::common::{
+    errors::WebsocketError,
     models::{ParamBuildError, WebsocketApiResponse},
     utils::remove_empty_value,
     websocket::{WebsocketApi, WebsocketMessageSendOptions},
@@ -55,6 +56,7 @@ pub trait TradeApi: Send + Sync {
     ) -> anyhow::Result<WebsocketApiResponse<Box<models::QueryOrderResponseResult>>>;
 }
 
+#[derive(Clone)]
 pub struct TradeApiClient {
     websocket_api_base: Arc<WebsocketApi>,
 }
@@ -846,12 +848,13 @@ impl TradeApi for TradeApiClient {
             .send_message::<Box<models::CancelOrderResponseResult>>(
                 "/order.cancel".trim_start_matches('/'),
                 payload,
-                WebsocketMessageSendOptions {
-                    is_signed: true,
-                    with_api_key: false,
-                },
+                WebsocketMessageSendOptions::new().signed(),
             )
             .await
+            .map_err(anyhow::Error::from)?
+            .into_iter()
+            .next()
+            .ok_or(WebsocketError::NoResponse)
             .map_err(anyhow::Error::from)
     }
 
@@ -897,12 +900,13 @@ impl TradeApi for TradeApiClient {
             .send_message::<Box<models::ModifyOrderResponseResult>>(
                 "/order.modify".trim_start_matches('/'),
                 payload,
-                WebsocketMessageSendOptions {
-                    is_signed: true,
-                    with_api_key: false,
-                },
+                WebsocketMessageSendOptions::new().signed(),
             )
             .await
+            .map_err(anyhow::Error::from)?
+            .into_iter()
+            .next()
+            .ok_or(WebsocketError::NoResponse)
             .map_err(anyhow::Error::from)
     }
 
@@ -1001,12 +1005,13 @@ impl TradeApi for TradeApiClient {
             .send_message::<Box<models::NewOrderResponseResult>>(
                 "/order.place".trim_start_matches('/'),
                 payload,
-                WebsocketMessageSendOptions {
-                    is_signed: true,
-                    with_api_key: false,
-                },
+                WebsocketMessageSendOptions::new().signed(),
             )
             .await
+            .map_err(anyhow::Error::from)?
+            .into_iter()
+            .next()
+            .ok_or(WebsocketError::NoResponse)
             .map_err(anyhow::Error::from)
     }
 
@@ -1037,12 +1042,13 @@ impl TradeApi for TradeApiClient {
             .send_message::<Vec<models::PositionInformationResponseResultInner>>(
                 "/account.position".trim_start_matches('/'),
                 payload,
-                WebsocketMessageSendOptions {
-                    is_signed: true,
-                    with_api_key: false,
-                },
+                WebsocketMessageSendOptions::new().signed(),
             )
             .await
+            .map_err(anyhow::Error::from)?
+            .into_iter()
+            .next()
+            .ok_or(WebsocketError::NoResponse)
             .map_err(anyhow::Error::from)
     }
 
@@ -1073,12 +1079,13 @@ impl TradeApi for TradeApiClient {
             .send_message::<Vec<models::PositionInformationV2ResponseResultInner>>(
                 "/v2/account.position".trim_start_matches('/'),
                 payload,
-                WebsocketMessageSendOptions {
-                    is_signed: true,
-                    with_api_key: false,
-                },
+                WebsocketMessageSendOptions::new().signed(),
             )
             .await
+            .map_err(anyhow::Error::from)?
+            .into_iter()
+            .next()
+            .ok_or(WebsocketError::NoResponse)
             .map_err(anyhow::Error::from)
     }
 
@@ -1114,12 +1121,13 @@ impl TradeApi for TradeApiClient {
             .send_message::<Box<models::QueryOrderResponseResult>>(
                 "/order.status".trim_start_matches('/'),
                 payload,
-                WebsocketMessageSendOptions {
-                    is_signed: true,
-                    with_api_key: false,
-                },
+                WebsocketMessageSendOptions::new().signed(),
             )
             .await
+            .map_err(anyhow::Error::from)?
+            .into_iter()
+            .next()
+            .ok_or(WebsocketError::NoResponse)
             .map_err(anyhow::Error::from)
     }
 }
@@ -1197,6 +1205,7 @@ mod tests {
             WebsocketHandler::on_message(&*ws_api, resp_json.to_string(), conn.clone()).await;
 
             let response = timeout(Duration::from_secs(1), handle).await.expect("task done").expect("no panic").expect("no error");
+
 
             let response_rate_limits = response.rate_limits.clone();
             let response_data = response.data().expect("deserialize data");
@@ -1327,6 +1336,7 @@ mod tests {
             WebsocketHandler::on_message(&*ws_api, resp_json.to_string(), conn.clone()).await;
 
             let response = timeout(Duration::from_secs(1), handle).await.expect("task done").expect("no panic").expect("no error");
+
 
             let response_rate_limits = response.rate_limits.clone();
             let response_data = response.data().expect("deserialize data");
@@ -1463,6 +1473,7 @@ mod tests {
 
             let response = timeout(Duration::from_secs(1), handle).await.expect("task done").expect("no panic").expect("no error");
 
+
             let response_rate_limits = response.rate_limits.clone();
             let response_data = response.data().expect("deserialize data");
 
@@ -1597,6 +1608,7 @@ mod tests {
 
             let response = timeout(Duration::from_secs(1), handle).await.expect("task done").expect("no panic").expect("no error");
 
+
             let response_rate_limits = response.rate_limits.clone();
             let response_data = response.data().expect("deserialize data");
 
@@ -1725,6 +1737,7 @@ mod tests {
 
             let response = timeout(Duration::from_secs(1), handle).await.expect("task done").expect("no panic").expect("no error");
 
+
             let response_rate_limits = response.rate_limits.clone();
             let response_data = response.data().expect("deserialize data");
 
@@ -1852,6 +1865,7 @@ mod tests {
             WebsocketHandler::on_message(&*ws_api, resp_json.to_string(), conn.clone()).await;
 
             let response = timeout(Duration::from_secs(1), handle).await.expect("task done").expect("no panic").expect("no error");
+
 
             let response_rate_limits = response.rate_limits.clone();
             let response_data = response.data().expect("deserialize data");